Add files via upload
Predict the cause of traffic accidents in UAE using a 2021 dataset and the accident size in UAE using a 2013-2019 dataset while taking other factors into account like weather condition, road condition and so on using a decision tree algorithm and done in Python. Two datasets have been used about the accidents that took place in UAE  from 2013-2019 as well as in 2021.